Treat signs differently to labels
Issues came about with scaling when labels & signs were treated the same
(it is possible to observe the scale of relations from a sign, but not
from a label). Commit explicitly separates them both in the spatial
layouts, abstract map, and visualiser.

Tidying up was also done with the designation of constraint sources (now
correctly determines whether the constraint came from a sign, label,
hierarchy, or unknown).